requirements.txt
Flask==3.0.0
Pillow==10.1.0
Werkzeug==3.0.1
requests==2.31.0